Make assembleLsstChannels copy all metadata
The raw images were coming in with minimal metadata,
so many fields needed for VisitInfo were missing.
Rather than specify the metadata keys fields to keep,
just accept them all (it would be easy to specify a list
of keys to omit, but I saw no need to do so).